refactor(chat): clean up types, remove redundancies and support AI flag
- Moved chat-related type definitions to `src/types/chat.ts`
- Reduced `Message` to essential fields (`text`, `label`, `type`, `sender`, optional `ai`, `buttons`)
- Removed redundant `clickable` and `source` properties
- Implemented parsing of `payload.ai === true` to mark AI-generated messages
- Adjusted mapping and component logic to reflect updated types
