+++ frontend/src/routes/trees/[treeId]/+page.svelte
@@ -4,12 +4,11 @@
 	import WaterColumn from '../../../components/WaterColumn.svelte';
 	import Chat from '../../../components/chat/Chat.svelte';
 	import Card from '../../../components/card/Card.svelte';
-	import Heading from '../../../components/typography/Heading.svelte';
 	import AdoptTree from '../../../features/adoption/AdoptTree.svelte';
 	import { supabase } from '../../../supabase';
 	import { onMount } from 'svelte';
 	import { page } from '$app/stores';
-	import type { Tree } from '../../../types/Tree';
+	import type { Tree } from '../../../types/tree';
 	import TreeMetric from '../../../components/trees/TreeMetric.svelte';
 
 	export let activeTabIndex: number = 0;
